+++ queue_services/entity-filer/src/entity_filer/filing_processors/conversion.py
@@ -28,7 +28,7 @@ from typing import Dict
 import dpath
 
 # from entity_queue_common.service_utils import BusinessException
-from business_model import db, AlternateName, LegalEntity, Filing
+from business_model import LegalEntity, Filing
 from entity_filer.utils.legislation_datetime import LegislationDatetime
 
 from entity_filer.exceptions import BusinessException, DefaultException
@@ -41,7 +41,8 @@ from entity_filer.filing_processors.filing_components import (
     shares,
 )
 from entity_filer.filing_processors.filing_components.offices import update_offices
-from entity_filer.filing_processors.filing_components.parties import get_or_create_party, merge_all_parties
+from entity_filer.filing_processors.filing_components.parties import merge_all_parties
+from entity_filer.filing_processors.filing_components.partner import update_partner_change, update_proprietor_change
 
 
 def process(
@@ -113,32 +114,20 @@ def _process_firms_conversion(
 ):
     match legal_entity.entity_type:
         case LegalEntity.EntityTypes.PARTNERSHIP:
-            _update_partner_change(
-                legal_entity,
-                filing_rec,
-                conversion_filing,
-                filing_meta
+            update_partner_change(
+                legal_entity=legal_entity,
+                filint_type="conversion",
+                change_filing_rec=filing_rec,
+                change_filing=conversion_filing,
+                filing_meta=filing_meta.conversion
             )
         case _: # LegalEntity.EntityTypes.PERSON: # legal_entity might be a proprietor?
-            _update_person_change(
-                legal_entity,
-                filing_rec,
-                conversion_filing,
-                filing_meta
+            update_proprietor_change(
+                filint_type="conversion",
+                change_filing_rec=filing_rec,
+                change_filing=conversion_filing,
+                filing_meta=filing_meta.conversion
             )
-
-    # Update Nature of Business
-    if (
-        naics := conversion_filing.get("filing", {})
-        .get("conversion", {})
-        .get("legal_entity", {})
-        .get("naics")
-    ) and naics.get("naicsDescription"):
-        legal_entity_info.update_naics_info(legal_entity, naics)
-        filing_meta.conversion = {
-            **filing_meta.conversion,
-            **{"naicsDescription": naics.get("naicsDescription")},
-        }
 
     # Update legal_entity office if present
     with suppress(IndexError, KeyError, TypeError):
@@ -161,164 +150,6 @@ def _process_firms_conversion(
                     legal_entity_start_date
                 )
             )
-
-
-def _update_partner_change(
-        legal_entity: LegalEntity,
-        change_filing_rec: Filing,
-        change_filing: Dict,
-        filing_meta: FilingMeta,
-):
-    name_request = dpath.util.get(change_filing, "/conversion/nameRequest", default=None)
-    if name_request and (to_legal_name := name_request.get("legalName")):
-        alternate_name = AlternateName.find_by_identifier(legal_entity.identifier)
-        parties_dict = dpath.util.get(change_filing, "/conversion/parties")
-
-        legal_entity.legal_name = get_partnership_name(parties_dict)
-
-        legal_entity.alternate_names.remove(alternate_name)
-        alternate_name.end_date = change_filing_rec.effective_date
-        alternate_name.change_filing_id = change_filing_rec.id
-        # alternate_name.delete()
-        db.session.add(alternate_name)
-        db.session.commit()
-        db.session.delete(alternate_name)
-        db.session.commit()
-
-
-        new_alternate_name = AlternateName(
-            bn15=alternate_name.bn15,
-            change_filing_id=change_filing_rec.id,
-            end_date=None,
-            identifier=legal_entity.identifier,
-            name=to_legal_name,
-            name_type=AlternateName.NameType.OPERATING,
-            start_date=change_filing.get("startDate"),
-            registration_date=change_filing_rec.effective_date,
-        )
-        legal_entity.alternate_names.append(new_alternate_name)
-
-        filing_meta.conversion = {
-            **filing_meta.conversion,
-            "fromLegalName": alternate_name.name,
-            "toLegalName": to_legal_name,
-        }
-
-    # Update Nature of LegalEntity
-    if (
-        naics := change_filing.get("conversion", {})
-        .get("business", {})
-        .get("naics")
-    ) and (naics_code := naics.get("naicsCode")):
-        if legal_entity.naics_code != naics_code:
-            filing_meta.conversion = {
-                **filing_meta.conversion,
-                **{
-                    "fromNaicsCode": legal_entity.naics_code,
-                    "toNaicsCode": naics_code,
-                    "naicsDescription": naics.get("naicsDescription"),
-                },
-            }
-            legal_entity_info.update_naics_info(legal_entity, naics)
-
-
-def _update_person_change(
-        legal_entity: LegalEntity,
-        change_filing_rec: Filing,
-        change_filing: Dict,
-        filing_meta: FilingMeta,
-):
-    name_request = dpath.util.get(change_filing, "/conversion/nameRequest", default=None)
-    identifier = dpath.util.get(change_filing_rec.filing_json, "filing/business/identifier")
-    if name_request and (to_legal_name := name_request.get("legalName")):
-        alternate_name = AlternateName.find_by_identifier(identifier)
-        parties_dict = dpath.util.get(change_filing, "/conversion/parties")
-
-        # Find the Proprietor
-        proprietor = None
-        for party in parties_dict:
-            for role in party.get("roles"):
-                if role.get("roleType") == "Proprietor":
-                    proprietor_dict = party
-                    break
-            if proprietor_dict:
-                break
-
-        if not proprietor_dict:
-            raise DefaultException(
-                f"No Proprietor in the SP conversion for filing:{change_filing_rec.id}"
-            )
-
-        proprietor, delivery_address, mailing_address = get_or_create_party(
-            proprietor_dict, change_filing_rec
-        )
-        if not proprietor:
-            raise DefaultException(
-                f"No Proprietor in the SP conversion for filing:{change_filing_rec.id}"
-            )
-        
-        if start := change_filing.get("filing", {}).get("conversion", {}).get("startDate"):
-            start_date = LegislationDatetime.as_utc_timezone_from_legislation_date_str(
-                start
-            )
-        elif change_filing.effective_date:
-            start_date = change_filing.effective_date.isoformat()
-        else:
-            start_date = LegislationDatetime.now()
-
-        alternate_name.end_date = change_filing_rec.effective_date
-        alternate_name.change_filing_id = change_filing_rec.id
-        # alternate_name.delete()
-        db.session.add(alternate_name)
-        db.session.commit()
-        db.session.delete(alternate_name)
-        db.session.commit()
-
-        new_alternate_name = AlternateName(
-            identifier=identifier,
-            name_type=AlternateName.NameType.OPERATING,
-            change_filing_id=change_filing_rec.id,
-            end_date=None,
-            name=to_legal_name,
-            start_date=start_date,
-            registration_date=change_filing_rec.effective_date,
-        )
-        proprietor.alternate_names.append(new_alternate_name)
-
-        filing_meta.conversion = {
-            **filing_meta.conversion,
-            "fromLegalName": alternate_name.name,
-            "toLegalName": to_legal_name,
-        }
-
-
-def get_partnership_name(parties_dict: dict):
-    """Set the legal_name of the partnership."""
-    parties = []
-    # get all parties in an array
-    for party in parties_dict:
-        if officer := party.get("officer"):
-            if org_name := officer.get("organizationName"):
-                parties.append(org_name.upper())
-                continue
-
-            name = officer["lastName"]
-            if first_name := officer.get("firstName"):
-                name = f"{name} {first_name}"
-            if middle_name := officer.get("middleName"):
-                name = f"{name} {middle_name}"
-            parties.append(name.upper())
-
-    if len(parties) < 2:
-        return parties[0]
-
-    parties.sort()
-    if parties and len(parties) > 2:
-        legal_name_str = ", ".join(parties[:2])
-        legal_name_str = f"{legal_name_str}, et al"
-    else:
-        legal_name_str = ", ".join(parties)
-    return legal_name_str
 
 
 def post_process(legal_entity: LegalEntity, filing: Filing):
